webserver/serverapp/requirements.txt
Django==1.10.6
PyJWT==1.4.2
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
gunicorn==19.7.0
pycrypto==2.6.1
requests==2.13.0
whitenoise==3.3.0
wsgiref==0.1.2
yagmail==0.6.159
